==> code/00_init.R <==
dir.create('./data')
dir.create('./data/grdc_raw')
dir.create('./data/grdc_reference_stations_raw')
dir.create('./results')
dir.create('./results/figures')
dir.create('./docs')
dir.create('./data/Selected_stations_raw_data')
dir.create('./data/Selected_stations_raw_data/Africa')
dir.create('./data/Selected_stations_raw_data/Asia')
dir.create('./data/Selected_stations_raw_data/Europe')
dir.create('./data/Selected_stations_raw_data/NorthCentralAmerica')
dir.create('./data/Selected_stations_raw_data/SouthAmerica')
dir.create('./data/Selected_stations_raw_data/SouthWestPacific')
dir.create('./data/Selected_stations_raw_data/Africa/Months')
dir.create('./data/Selected_stations_raw_data/Africa/Days')
dir.create('./data/Selected_stations_raw_data/Asia/Months')
dir.create('./data/Selected_stations_raw_data/Asia/Days')
dir.create('./data/Selected_stations_raw_data/Europe/Months')
dir.create('./data/Selected_stations_raw_data/Europe/Days')
dir.create('./data/Selected_stations_raw_data/NorthCentralAmerica/Months')
dir.create('./data/Selected_stations_raw_data/NorthCentralAmerica/Days')
dir.create('./data/Selected_stations_raw_data/SouthAmerica/Months')
dir.create('./data/Selected_stations_raw_data/SouthAmerica/Days')
dir.create('./data/Selected_stations_raw_data/SouthWestPacific/Months')
dir.create('./data/Selected_stations_raw_data/SouthWestPacific/Days')
